Removed AppInfo and AppConfirmation
